refactor(workbench): extract center chrome into WorkbenchMain
Title row, toast, drop overlay, and top actions leave AppWorkbench. Chat, kanban, and automations stay as children.
